pyproject.toml
[project]
name = "fastapi_openai_mcp"
version = "0.1.0"
description = "Sample FastAPI project integrating OpenAI tool calling with an MCP service"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "python-dotenv",
    "openai",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "coverage"]

[tool.pytest.ini_options]
addopts = "-q"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
